test: Improvements for builder pattern
