+++ crates/solver-config/src/builders/mod.rs
@@ -1,0 +1,5 @@
+//! Configuration builders for test and development scenarios.
+
+pub mod config;
+
+pub use config::ConfigBuilder;
