+++ crates/solver-service/Cargo.toml
@@ -45,4 +45,3 @@ once_cell = "1.20"
 tempfile = { workspace = true }
 tokio-test = "0.4"
 mockall = "0.12"
-solver-config = { path = "../solver-config", features = ["testing"] }
